# server/docs/swagger.yaml
basePath: /
definitions:
  dto.ChatHistoryItem:
    properties:
      created_at:
        type: integer
      id:
        type: string
      question:
        type: string
      updated_at:
        type: integer
    type: object
...
  dto.CreateChatRequest:
    properties:
      message:
        example: Who are you having lasagna with today and why?
        type: string
    required:
    - message
    type: object
  dto.CreateChatResponse:
    properties:
      created_at:
        type: integer
      id:
        type: string
      message:
        type: string
      response:
        type: string
    type: object
...
  dto.GetChatResponse:
    properties:
      created_at:
        type: integer
      id:
        type: string
      messages:
        items:
          additionalProperties: true
          type: object
        type: array
      updated_at:
        type: integer
    type: object
...
  dto.GetUserHistoryResponse:
    properties:
      chats:
        items:
          $ref: '#/definitions/dto.ChatHistoryItem'
        type: array
      nextCursor:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        example: msk.vitaly@gmail.com
        type: string
      organization:
        example: Staffbase
        type: string
      password:
        example: defaultPassword
        type: string
    required:
    - email
    - organization
    - password
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  dto.SuccessLogoutResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.UnauthorizedResponse:
    properties:
      message:
        type: string
    type: object
...
  jwtUtils.UserClaims:
    properties:
      _id:
        type: string
      organizationId:
        type: string
      organizationRole:
        type: string
    type: object
  lib.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact:
    email: msk.vitaly@gmail.com
    name: Vitalii Popov
    url: https://www.linkedin.com/in/mskVitalii/
  description: Semantic Contacts
  title: Semki
  version: "1.0"
paths:
  /api/v1/claims:
    get:
      description: Fetches and returns the claims of the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with user claims
          schema:
            $ref: '#/definitions/jwtUtils.UserClaims'
        "401":
          description: Invalid claims type
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
      security:
      - BearerAuth: []
      summary: Retrieves user claims
      tags:
      - auth
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password, and returns a JWT
        token if the credentials are valid.
      parameters:
      - description: Login request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
      summary: Authenticates a user and returns a JWT token
      tags:
      - auth
  /api/v1/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the JWT token for the user by adding it to the blacklist
        and removes the JWT cookie.
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successful logout
          schema:
            $ref: '#/definitions/dto.SuccessLogoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logs out a user and invalidates the JWT token
      tags:
      - auth
...
  /api/v1/refresh_token:
    post:
      consumes:
      - application/json
      description: Generates a new authentication token using the refresh token provided.
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.UnauthorizedResponse'
      security:
      - BearerAuth: []
      summary: Refreshes the authentication token
      tags:
      - auth
...
  /chat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Chat creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new chat
      tags:
      - chats
  /chat/{id}:
    get:
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get chat by ID
      tags:
      - chats
  /chat/history:
    get:
      parameters:
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      - default: 20
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetUserHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user chat history
      tags:
      - chats
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"