# server/pkg/lib/requests.go
package lib

import (
	"github.com/getsentry/sentry-go"
	"github.com/gin-gonic/gin"
	"net/http"
	"semki/pkg/telemetry"
	"time"
)

type ErrorResponse struct {
	Message string `json:"message"`
}

func ResponseBadRequest(c *gin.Context, err error, message string) {
	sentry.CaptureException(err)
	sentry.Flush(time.Second * 5)
	telemetry.Log.Error(message)
	c.JSON(http.StatusBadRequest, ErrorResponse{Message: message})
}

func ResponseInternalServerError(c *gin.Context, err error, message string) {
	sentry.CaptureException(err)
	sentry.Flush(time.Second * 5)
	telemetry.Log.Error(message)
	c.JSON(http.StatusInternalServerError, ErrorResponse{Message: message})
}

func ResponseNotFound(c *gin.Context, message string) {
	c.JSON(http.StatusNotFound, ErrorResponse{Message: message})
}

func Preflight(c *gin.Context) {
	c.Header("Access-Control-Allow-Origin", "*")
	c.Header("Access-Control-Allow-Methods", "GET, OPTIONS")
	c.Header("Access-Control-Allow-Headers", "Content-Type, Authorization")
	c.Status(http.StatusOK)
}